Ensure that our container div owns the entire browser window
Because it's basically a window manager, the Tracery UI is a pixel-level
UI rather than being flexbox or whatnot. This means we need to know the parent's
size in pixels at all times, and we can't just style the app to fill the window.

This commit sets up the necessary infrastructure for that.