Fix how ESLint works with Flow type vars
In addition to letting you annotate types in your JavaScript, Flow also has a couple ways of defining types that don't exist in the JavaScript code. There are two common cases for this:
1. Defining shapes of objects (e.g. React Props objects)
2. Type declarations for code that's not written in Flow

Stock ESLint is confused by these things and reports undefined symbol errors all over the place. This change brings in a handy plugin that makes it be quiet.