Fix ESLint config to allow `import type`
Flow allows you to import type declarations from another file with `import type`. However, if you import both Flow type declarations and regular JS declarations from the same file, you wind up with both an `import` and an `import type`, and ESLint thinks this is a duplicate. The `import` plugin for ESLint is smarter, so switching to use that for this particular rule.