Use Armeria for server
Thrift's generated JS code can either use XMLHttpRequest or WebSockets to talk to its server. Unfortunately the default Thrift Java server supports neither of those. We found this Armeria project that does support at least XHR out of the box.

Unfortunately, it does not appear to support CORS headers, so for now it is set up to serve our built webpack bundle as well (and we'll need to hit localhost:9090 in the browser). This slows UI dev iteration time something fierce, so we need to find a way to get the appropriate headers in there. It sounds like webpack-dev-server is supposed to be able to proxy requests to a different backend, but I haven't been able to get that to work yet.