Clean up state helpers
We were doing a thing where we were writing helper functions all of the place.
That's nice for reducing the number of places that need to change when the
layout of the state changes

That's especially important if you have to actually run the code in order to
discover if you've missed anything. However, we use Flow typing throughout
the state, so we can save the typing, reach in from the global level, and
know that Flow will find all the places we need to change when structure
changes.