TraceList -> QueryBackedList
Just like we generalized the state and actions around querying, we can also generalize the UI.
This commit takes the short hop from `TraceList` as a list of traces to `QueryBackedList` as a
list of any query result.

This is still a very dumb component, and will eventually be replaced with something much nicer.